pyproject.toml
[build-system]
requires = ["setuptools>=61.2"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["xsdata_attrs*"]

[project]
name = "xsdata_attrs"
description = "xsdata attrs plugin"
authors = [{name = "Christodoulos Tsoulloftas", email = "chris@komposta.net"}]
license = {text = "MIT"}
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
    "Topic :: Software Development :: Code Generators",
    "Topic :: Text Processing :: Markup :: XML",
]
keywords = ["xsd", "wsdl", "schema", "dtd", "binding", "xml", "json", "dataclasses", "generator", "cli", "attrs"]
requires-python = ">=3.8"
dependencies = [
    "attrs>=19.2.0",
    "xsdata>=24.3",
]
dynamic = ["version"]

[project.readme]
file = "README.md"
content-type = "text/markdown"

[project.urls]
Homepage = "https://github.com/tefra/xsdata-attrs"
Source = "https://github.com/tefra/xsdata-attrs"
Documentation = "https://xsdata-attrs.readthedocs.io/"
Changelog = "https://xsdata-attrs.readthedocs.io/en/latest/changelog/"

[project.entry-points]
"xsdata.plugins.cli" = {xsdata_attrs = "xsdata_attrs.hooks.cli"}

[project.entry-points."xsdata.plugins.class_types"]
xsdata_attrs = "xsdata_attrs.hooks.class_type"

[project.optional-dependencies]
cli = ["xsdata[cli]>=23.5"]
docs = [
    "mkdocs",
    "mkdocs-gen-files",
    "mkdocs-literate-nav",
    "mkdocs-material",
    "mkdocs-minify-plugin",
    "mkdocstrings[python]",
    "markdown-exec[ansi]",
    "pymdownx-superfence-filter-lines",
]
lxml = ["lxml>=4.5.0"]
soap = ["requests"]
test = [
    "codecov",
    "lxml",
    "pre-commit",
    "pytest",
    "pytest-cov",
]

[tool.setuptools]
include-package-data = true
license-files = ["LICENSE"]

[tool.setuptools.dynamic]
version = {attr = "xsdata_attrs.__version__"}

[tool.doc8]
max-line-length = 88

[tool.pytest.ini_options]
addopts = "--color=yes"
